ultima clase navidad

+++ freeTime/src/main/java/com/example/demo/repository/OpinionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.example.demo.entity.Opinion;
 
-public interface OpinionRepository extends CrudRepository<Opinion, Integer> {
+public interface OpinionRepository extends CrudRepository<Opinion, Long> {
 
 }
